backend/requirements.txt
# OMEN Backend Dependencies
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Validation & Settings
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Redis
redis[hiredis]==5.2.1

# HTTP Client
httpx==0.28.1
aiohttp==3.11.11

# Stripe
stripe==11.4.1

# Logging
structlog==24.4.0

# Utilities
python-dateutil==2.9.0
orjson==3.10.12
